Translates logit regression coefficients into percentages

Performs a logit regression and then computes the effects of covariates expressed in percentages (through two methods: 'pure' effects and 'experimental' effects; see Deauvieau, 2010)

translate.logit(formula,data,nit=0)

Arguments

formula

an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Every variables have to be factors.

data

a data frame containing the variables in the model

nit

number of bootstrap iterations for confidence interval computation. Default is 0, i.e. no confidence interval is computed.

Details

This function works with binomial as well as multinomial regression models. If the dependant variable has two factors, glm is used, if it has more than two factors multinom function (from nnet package) is used. The function expresses the regression coefficients as percentages through three distinct methods: raw percentages, 'pure effects' percentages and 'experimental effects' percentages (see Deauvieau, 2010). Bootstrap confidence interval are available only for binomial regressions.

Value

The function returns a list:

glm

An object of class glm or nnet (depending on the number of factors of the dependent variable)

summary

The results of summary function applied to reg element

percents

A matrix or a list of matrices (depending on the number of factors of the dependent variable) with regression coefficients expressed as percentages

boot.ci

A matrix or a list of matrices (depending on the number of factors of the dependent variable) with confidence intervals computed with bootstrap

References

Deauvieau, J. (2010), 'Comment traduire sous forme de probabilites les resultats d'une modelisation logit ?', Bulletin of Sociological Methodology / Bulletin de Methodologie Sociologique 105(1), 5-23.

Deauvieau, J. (2011), 'Est-il possible et souhaitable traduire sous forme de probabilites un coefficient logit ? Reponse aux remarques formulees par Marion Selz a propos de mon article paru dans le BMS en 2010', Bulletin of Sociological Methodology / Bulletin de Methodologie Sociologique 112(1), 32-42.

Deauvieau, J. (2019), 'Comparer les resultats d’un modele logit dichotomique ou polytomique entre plusieurs groupes a partir des probabilites estimees', Bulletin of Sociological Methodology / Bulletin de Methodologie Sociologique 142(1), 7-31.

Author

Nicolas Robette

See also

Examples

  ## An example for binomial logit regression
  data(Movies)
  translate.logit(ArtHouse ~ Genre + Festival, Movies)
#> $glm
#> 
#> Call:  glm(formula = f, family = "binomial", data = d)
#> 
#> Coefficients:
#>      (Intercept)    GenreAnimation        GenreOther      GenreComDram  
#>          -1.3024           -0.3330            0.7573            1.8669  
#>      GenreComedy  GenreDocumentary        GenreDrama       GenreHorror  
#>           0.2604            3.2982            2.4892          -15.2636  
#>       GenreSciFi       FestivalYes  
#>          -1.1586            2.3045  
#> 
#> Degrees of Freedom: 999 Total (i.e. Null);  990 Residual
#> Null Deviance:	    1386 
#> Residual Deviance: 1002 	AIC: 1022
#> 
#> $summary
#> 
#> Call:
#> glm(formula = f, family = "binomial", data = d)
#> 
#> Deviance Residuals: 
#>      Min        1Q    Median        3Q       Max  
#> -2.65379  -0.77734  -0.00036   0.72987   2.25519  
#> 
#> Coefficients:
#>                  Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)       -1.3024     0.1922  -6.778 1.22e-11 ***
#> GenreAnimation    -0.3330     0.4395  -0.758   0.4486    
#> GenreOther         0.7573     0.4589   1.650   0.0989 .  
#> GenreComDram       1.8669     0.2599   7.183 6.82e-13 ***
#> GenreComedy        0.2604     0.2444   1.065   0.2867    
#> GenreDocumentary   3.2982     0.4037   8.170 3.10e-16 ***
#> GenreDrama         2.4892     0.2504   9.939  < 2e-16 ***
#> GenreHorror      -15.2636   479.9090  -0.032   0.9746    
#> GenreSciFi        -1.1586     0.5312  -2.181   0.0292 *  
#> FestivalYes        2.3045     0.4022   5.730 1.00e-08 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 1385.5  on 999  degrees of freedom
#> Residual deviance: 1002.0  on 990  degrees of freedom
#> AIC: 1022
#> 
#> Number of Fisher Scoring iterations: 15
#> 
#> 
#> $percents
#>                        raw     expe     pure expe.lin
#> Genre.Action      0.254545 0.253620 0.238621 0.253967
#> Genre.Animation   0.173913 0.201444 0.183428 0.191778
#> Genre.Other       0.423077 0.404417 0.400591 0.410668
#> Genre.ComDram     0.664430 0.661265 0.669660 0.661116
#> Genre.Comedy      0.274775 0.300695 0.289077 0.290930
#> Genre.Documentary 0.883117 0.888539 0.894541 0.899612
#> Genre.Drama       0.796680 0.781896 0.790670 0.773282
#> Genre.Horror      0.000000 0.000000 0.000000 0.024892
#> Genre.SciFi       0.102041 0.108072 0.089572 0.107141
#> Festival.No       0.452871 0.461473 0.452122 0.461108
#> Festival.Yes      0.883117 0.819758 0.892102 0.784386
#> 
  
  ## An example for multinomial logit regression
  translate.logit(Country ~ Genre + Festival, Movies)
#> # weights:  44 (30 variable)
#> initial  value 1386.294361 
#> iter  10 value 879.086422
#> iter  20 value 828.776089
#> iter  30 value 827.782446
#> iter  40 value 827.747080
#> final  value 827.744631 
#> converged
#> $glm
#> Call:
#> multinom(formula = formula, data = data, model = TRUE)
#> 
#> Coefficients:
#>        (Intercept) GenreAnimation  GenreOther GenreComDram GenreComedy
#> France   1.4560336      0.7462496  -0.2523886    0.9611346   0.7472260
#> Other   -0.8170293    -11.7578656 -51.3421343   -0.3537963  -0.5589406
#> USA      1.8357724      0.7349839  -0.9956849   -0.8577987  -0.5409056
#>        GenreDocumentary GenreDrama GenreHorror GenreSciFi FestivalYes
#> France        2.8407089  0.8737679 -2.55457635 -0.7490637  -0.2021490
#> Other         0.8276979  0.1805444 -0.28238140  0.1475178  -0.3749129
#> USA          -1.1352225 -1.1325203  0.06116968  0.3493737  -0.2396825
#> 
#> Residual Deviance: 1655.489 
#> AIC: 1715.489 
#> 
#> $summary
#> Call:
#> multinom(formula = formula, data = data, model = TRUE)
#> 
#> Coefficients:
#>        (Intercept) GenreAnimation  GenreOther GenreComDram GenreComedy
#> France   1.4560336      0.7462496  -0.2523886    0.9611346   0.7472260
#> Other   -0.8170293    -11.7578656 -51.3421343   -0.3537963  -0.5589406
#> USA      1.8357724      0.7349839  -0.9956849   -0.8577987  -0.5409056
#>        GenreDocumentary GenreDrama GenreHorror GenreSciFi FestivalYes
#> France        2.8407089  0.8737679 -2.55457635 -0.7490637  -0.2021490
#> Other         0.8276979  0.1805444 -0.28238140  0.1475178  -0.3749129
#> USA          -1.1352225 -1.1325203  0.06116968  0.3493737  -0.2396825
#> 
#> Std. Errors:
#>        (Intercept) GenreAnimation   GenreOther GenreComDram GenreComedy
#> France   0.3000539      0.8030366 6.447938e-01    0.4444493   0.3981802
#> Other    0.4921245    381.8897013 5.125651e-08    0.8195146   0.7432585
#> USA      0.2915326      0.7890573 6.668749e-01    0.4707400   0.4045597
#>        GenreDocumentary GenreDrama GenreHorror GenreSciFi FestivalYes
#> France         1.050384  0.3883454   1.1929239  0.6808290   0.4370104
#> Other          1.496746  0.6387992   1.2554242  0.9942557   0.8503507
#> USA            1.258719  0.4105479   0.6843009  0.6015092   0.4780629
#> 
#> Residual Deviance: 1655.489 
#> AIC: 1715.489 
#> 
#> $percents
#> $percents$by_category
#> $percents$by_category$Europe
#>                         raw      expe      pure  expe.lin
#> Genre.Action      -0.968219 -0.969055 -0.967064 -0.969090
#> Genre.Animation   -0.828571 -0.826907 -0.826745 -0.816462
#> Genre.Other       -0.457014 -0.458311 -0.457783 -0.447752
#> Genre.ComDram     -0.869658 -0.868782 -0.866775 -0.868433
#> Genre.Comedy      -0.883784 -0.875037 -0.873154 -0.871812
#> Genre.Documentary -1.153153 -1.146683 -1.145417 -1.146835
#> Genre.Drama       -0.902339 -0.917359 -0.915375 -0.916291
#> Genre.Horror      -0.369565 -0.357875 -0.355606 -0.356790
#> Genre.SciFi       -0.897436 -0.905565 -0.903545 -0.904758
#> Festival.No       -0.975489 -0.976670 -0.975297 -0.976608
#> Festival.Yes      -0.827991 -0.851027 -0.844782 -0.831399
#> 
#> $percents$by_category$France
#>                        raw     expe     pure expe.lin
#> Genre.Action      0.808219 0.808204 0.808043 0.808210
#> Genre.Animation   0.900000 0.898586 0.898513 0.898538
#> Genre.Other       0.764706 0.765857 0.765652 0.765323
#> Genre.ComDram     0.916667 0.916795 0.916738 0.916825
#> Genre.Comedy      0.900000 0.899124 0.899052 0.899090
#> Genre.Documentary 0.986486 0.986309 0.986301 0.985263
#> Genre.Drama       0.908629 0.909704 0.909640 0.909859
#> Genre.Horror      0.250000 0.247377 0.246934 0.248538
#> Genre.SciFi       0.666667 0.666691 0.666381 0.666678
#> Festival.No       0.895330 0.895106 0.895122 0.895111
#> Festival.Yes      0.875000 0.878439 0.877303 0.877431
#> 
#> $percents$by_category$Other
#>                        raw     expe     pure  expe.lin
#> Genre.Action      0.300000 0.300857 0.299194  0.300889
#> Genre.Animation   0.000000 0.000000 0.000000 -0.009999
#> Genre.Other       0.000000 0.000000 0.000000 -0.009999
#> Genre.ComDram     0.230769 0.229491 0.227839  0.229146
#> Genre.Comedy      0.200000 0.193164 0.191599  0.190001
#> Genre.Documentary 0.500000 0.487585 0.486664  0.490001
#> Genre.Drama       0.333333 0.344043 0.342465  0.343497
#> Genre.Horror      0.250000 0.241798 0.240130  0.240001
#> Genre.SciFi       0.333333 0.341885 0.340301  0.341480
#> Festival.No       0.269663 0.275343 0.274112  0.275305
#> Festival.Yes      0.222222 0.182342 0.178231  0.166430
#> 
#> $percents$by_category$USA
#>                        raw     expe     pure expe.lin
#> Genre.Action      0.860000 0.859994 0.859827 0.859991
#> Genre.Animation   0.928571 0.928321 0.928232 0.927923
#> Genre.Other       0.692308 0.692454 0.692131 0.692428
#> Genre.ComDram     0.722222 0.722496 0.722198 0.722462
#> Genre.Comedy      0.783784 0.782749 0.782503 0.782721
#> Genre.Documentary 0.666667 0.672789 0.672452 0.671571
#> Genre.Drama       0.660377 0.663612 0.663270 0.662935
#> Genre.Horror      0.869565 0.868700 0.868542 0.868251
#> Genre.SciFi       0.897436 0.896989 0.896863 0.896600
#> Festival.No       0.810496 0.806221 0.806063 0.806192
#> Festival.Yes      0.730769 0.790246 0.789248 0.787538
#> 
#> 
#> $percents$by_method
#> $percents$by_method$raw
#>                      Europe   France    Other      USA
#> Genre.Action      -0.968219 0.808219 0.300000 0.860000
#> Genre.Animation   -0.828571 0.900000 0.000000 0.928571
#> Genre.Other       -0.457014 0.764706 0.000000 0.692308
#> Genre.ComDram     -0.869658 0.916667 0.230769 0.722222
#> Genre.Comedy      -0.883784 0.900000 0.200000 0.783784
#> Genre.Documentary -1.153153 0.986486 0.500000 0.666667
#> Genre.Drama       -0.902339 0.908629 0.333333 0.660377
#> Genre.Horror      -0.369565 0.250000 0.250000 0.869565
#> Genre.SciFi       -0.897436 0.666667 0.333333 0.897436
#> Festival.No       -0.975489 0.895330 0.269663 0.810496
#> Festival.Yes      -0.827991 0.875000 0.222222 0.730769
#> 
#> $percents$by_method$expe
#>                      Europe   France    Other      USA
#> Genre.Action      -0.969055 0.808204 0.300857 0.859994
#> Genre.Animation   -0.826907 0.898586 0.000000 0.928321
#> Genre.Other       -0.458311 0.765857 0.000000 0.692454
#> Genre.ComDram     -0.868782 0.916795 0.229491 0.722496
#> Genre.Comedy      -0.875037 0.899124 0.193164 0.782749
#> Genre.Documentary -1.146683 0.986309 0.487585 0.672789
#> Genre.Drama       -0.917359 0.909704 0.344043 0.663612
#> Genre.Horror      -0.357875 0.247377 0.241798 0.868700
#> Genre.SciFi       -0.905565 0.666691 0.341885 0.896989
#> Festival.No       -0.976670 0.895106 0.275343 0.806221
#> Festival.Yes      -0.851027 0.878439 0.182342 0.790246
#> 
#> $percents$by_method$pure
#>                      Europe   France    Other      USA
#> Genre.Action      -0.967064 0.808043 0.299194 0.859827
#> Genre.Animation   -0.826745 0.898513 0.000000 0.928232
#> Genre.Other       -0.457783 0.765652 0.000000 0.692131
#> Genre.ComDram     -0.866775 0.916738 0.227839 0.722198
#> Genre.Comedy      -0.873154 0.899052 0.191599 0.782503
#> Genre.Documentary -1.145417 0.986301 0.486664 0.672452
#> Genre.Drama       -0.915375 0.909640 0.342465 0.663270
#> Genre.Horror      -0.355606 0.246934 0.240130 0.868542
#> Genre.SciFi       -0.903545 0.666381 0.340301 0.896863
#> Festival.No       -0.975297 0.895122 0.274112 0.806063
#> Festival.Yes      -0.844782 0.877303 0.178231 0.789248
#> 
#> $percents$by_method$expe.lin
#>                      Europe   France     Other      USA
#> Genre.Action      -0.969090 0.808210  0.300889 0.859991
#> Genre.Animation   -0.816462 0.898538 -0.009999 0.927923
#> Genre.Other       -0.447752 0.765323 -0.009999 0.692428
#> Genre.ComDram     -0.868433 0.916825  0.229146 0.722462
#> Genre.Comedy      -0.871812 0.899090  0.190001 0.782721
#> Genre.Documentary -1.146835 0.985263  0.490001 0.671571
#> Genre.Drama       -0.916291 0.909859  0.343497 0.662935
#> Genre.Horror      -0.356790 0.248538  0.240001 0.868251
#> Genre.SciFi       -0.904758 0.666678  0.341480 0.896600
#> Festival.No       -0.976608 0.895111  0.275305 0.806192
#> Festival.Yes      -0.831399 0.877431  0.166430 0.787538
#> 
#> 
#>